CAD-2221 locli:  factor modules
