+++ locli/locli.cabal
@@ -23,11 +23,14 @@ library
                        Cardano.TopHandler
 
                        Cardano.Unlog.Analysis
+                       Cardano.Unlog.ChainParams
                        Cardano.Unlog.Commands
                        Cardano.Unlog.LogObject
                        Cardano.Unlog.Parsers
                        Cardano.Unlog.Resources
                        Cardano.Unlog.Run
+                       Cardano.Unlog.SlotStats
+                       Cardano.Unlog.Summary
 
   other-modules:       Paths_locli
