+++ locli/src/Cardano/Unlog/Run.hs
@@ -9,12 +9,11 @@ module Cardano.Unlog.Run
 
 import           Cardano.Prelude
 
-import           Control.Monad.Trans.Except (ExceptT)
 import           Control.Monad.Trans.Except.Extra (firstExceptT)
 import qualified Data.Text as Text
 
 import           Cardano.Unlog.Commands (AnalysisCommand)
-import           Cardano.Unlog.Analysis (AnalysisCmdError, renderAnalysisCmdError,
+import           Cardano.Unlog.Summary (AnalysisCmdError, renderAnalysisCmdError,
                      runAnalysisCommand)
 
 import           Cardano.Config.Git.Rev (gitRev)
